mam2: extract and bazelify tests
