+++ mam/v2/mam/BUILD
@@ -1,0 +1,14 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "mam",
+    srcs = ["mam.c"],
+    hdrs = ["mam.h"],
+    deps = [
+        "//mam/v2:defs",
+        "//mam/v2:list",
+        "//mam/v2/ntru",
+        "//mam/v2/pb3",
+        "//mam/v2/prng",
+    ],
+)
