+++ mam/v2/mam/mam.h
@@ -296,7 +296,18 @@ typedef struct _mam_recv_packet_context {
 err_t mam_recv_packet(mam_recv_packet_context *cfg, trits_t *packet,
                       trits_t *payload);
 
-bool_t mam_test(isponge *s, void *sponge_alloc_ctx,
-                isponge *(create_sponge)(void *ctx),
-                void (*destroy_sponge)(void *ctx, isponge *), iprng *pa,
-                iprng *pb);
+trits_t mam_send_msg_cfg_chid(mam_send_msg_context *cfg);
+trits_t mam_send_msg_cfg_chid1(mam_send_msg_context *cfg);
+trits_t mam_send_msg_cfg_epid(mam_send_msg_context *cfg);
+trits_t mam_send_msg_cfg_epid1(mam_send_msg_context *cfg);
+trits_t mam_send_msg_cfg_nonce(mam_send_msg_context *cfg);
+trits_t mam_send_msg_cfg_key(mam_send_msg_context *cfg);
+
+trits_t mam_recv_msg_cfg_chid(mam_recv_msg_context *cfg);
+trits_t mam_recv_msg_cfg_chid1(mam_recv_msg_context *cfg);
+trits_t mam_recv_msg_cfg_epid(mam_recv_msg_context *cfg);
+trits_t mam_recv_msg_cfg_epid1(mam_recv_msg_context *cfg);
+trits_t mam_recv_msg_cfg_nonce(mam_recv_msg_context *cfg);
+trits_t mam_recv_msg_cfg_key(mam_recv_msg_context *cfg);
+trits_t mam_recv_msg_cfg_psk_id(mam_recv_msg_context *cfg);
+trits_t mam_recv_msg_cfg_ntru_id(mam_recv_msg_context *cfg);
