+++ mam/v2/mam/tests/BUILD
@@ -1,0 +1,8 @@
+cc_test(
+    name = "test_mam",
+    srcs = ["test_mam.c"],
+    deps = [
+        "//mam/v2/mam",
+        "@unity",
+    ],
+)
