+++ mam/v2/mss/tests/BUILD
@@ -1,0 +1,8 @@
+cc_test(
+    name = "test_mss",
+    srcs = ["test_mss.c"],
+    deps = [
+        "//mam/v2/mss",
+        "@unity",
+    ],
+)
