+++ mam/v2/ntru/tests/BUILD
@@ -1,0 +1,17 @@
+cc_test(
+    name = "test_ntru",
+    srcs = ["test_ntru.c"],
+    deps = [
+        "//mam/v2/ntru",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "test_poly",
+    srcs = ["test_poly.c"],
+    deps = [
+        "//mam/v2/ntru:poly",
+        "@unity",
+    ],
+)
