+++ mam/v2/pb3/tests/BUILD
@@ -1,0 +1,8 @@
+cc_test(
+    name = "test_pb3",
+    srcs = ["test_pb3.c"],
+    deps = [
+        "//mam/v2/pb3",
+        "@unity",
+    ],
+)
