+++ mam/v2/prng/tests/BUILD
@@ -1,0 +1,8 @@
+cc_test(
+    name = "test_prng",
+    srcs = ["test_prng.c"],
+    deps = [
+        "//mam/v2/prng",
+        "@unity",
+    ],
+)
