+++ mam/v2/sponge/tests/BUILD
@@ -1,0 +1,17 @@
+cc_test(
+    name = "test_sponge",
+    srcs = ["test_sponge.c"],
+    deps = [
+        "//mam/v2/sponge",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "test_spongos",
+    srcs = ["test_spongos.c"],
+    deps = [
+        "//mam/v2/sponge:spongos",
+        "@unity",
+    ],
+)
