+++ mam/v2/trits/tests/BUILD
@@ -1,0 +1,8 @@
+cc_test(
+    name = "test_trits",
+    srcs = ["test_trits.c"],
+    deps = [
+        "//mam/v2/trits",
+        "@unity",
+    ],
+)
