+++ mam/v2/wots/tests/BUILD
@@ -1,0 +1,8 @@
+cc_test(
+    name = "test_wots",
+    srcs = ["test_wots.c"],
+    deps = [
+        "//mam/v2/wots",
+        "@unity",
+    ],
+)
